Added user profile edit Major fixes

+++ src/Indigo/ContestBundle/Controller/ContestController.php
@@ -16,7 +16,6 @@ use Indigo\ContestBundle\Form\ContestType;
 
 class ContestController extends Controller
 {
-
     /**
      * Lists all Contest entities.
      *
@@ -226,7 +225,7 @@ class ContestController extends Controller
             $entity->uploadPrizeImage();
             $em->flush();
 
-            return $this->redirect($this->generateUrl('contest_edit', array('id' => $id)));
+            return $this->redirect($this->generateUrl('contest'));
         }
 
         return array(
@@ -244,15 +243,17 @@ class ContestController extends Controller
      */
     public function deleteAction(Request $request, $id)
     {
-        if ($request->isMethod("DELETE")) {
-            $form = $this->createDeleteForm($id);
-            $form->handleRequest($request);
+        if (Contest::OPEN_CONTEST_ID != $id){
+            if ($request->isMethod("DELETE")) {
+                $form = $this->createDeleteForm($id);
+                $form->handleRequest($request);
 
-            if ($form->isValid()) {
+                if ($form->isValid()) {
+                    $this->deleteItem($id);
+                }
+            } else {
                 $this->deleteItem($id);
             }
-        } else {
-            $this->deleteItem($id);
         }
 
         return $this->redirect($this->generateUrl('contest'));
